refactor: 기존 보드 수정 modal 리팩토링
